+++ packages/devtools_app/lib/devtools_app.dart
@@ -23,7 +23,7 @@ export 'src/screens/debugger/syntax_highlighter.dart';
 export 'src/screens/deep_link_validation/deep_links_controller.dart';
 export 'src/screens/deep_link_validation/deep_links_screen.dart';
 export 'src/screens/inspector/inspector_controller.dart';
-export 'src/screens/inspector/inspector_screen.dart';
+export 'src/screens/inspector/inspector_screen_body.dart';
 export 'src/screens/inspector/inspector_tree_controller.dart';
 export 'src/screens/logging/logging_controller.dart';
 export 'src/screens/logging/logging_screen.dart';
