+++ packages/devtools_app/lib/src/screens/inspector/inspector_screen_body.dart
