+++ packages/devtools_app/lib/src/screens/inspector_v2/inspector_controls.dart
@@ -14,7 +14,7 @@ import '../../shared/common_widgets.dart';
 import '../../shared/globals.dart';
 import '../inspector_shared/inspector_settings_dialog.dart';
 import 'inspector_controller.dart';
-import 'inspector_screen.dart';
+import 'inspector_screen_body.dart';
 
 /// Control buttons for the inspector panel.
 class InspectorControls extends StatelessWidget {
