+++ packages/devtools_app/test/inspector/inspector_tree_test.dart
@@ -60,6 +60,7 @@ void main() {
         inspector: inspectorController,
         debugger: debuggerController,
         InspectorTree(
+          controller: inspectorController,
           treeController: treeController,
           summaryTreeController: summaryTreeController,
           isSummaryTree: isSummaryTree,
