+++ packages/devtools_app/test/inspector_v2/inspector_integration_test.dart
@@ -3,8 +3,9 @@
 // found in the LICENSE file.
 
 import 'package:devtools_app/devtools_app.dart'
-    hide InspectorScreen, InspectorScreenBodyState, InspectorScreenBody;
-import 'package:devtools_app/src/screens/inspector_v2/inspector_screen.dart';
+    hide InspectorScreenBodyState, InspectorScreenBody;
+import 'package:devtools_app/src/screens/inspector_shared/inspector_screen.dart';
+import 'package:devtools_app/src/screens/inspector_v2/inspector_screen_body.dart';
 import 'package:devtools_app/src/screens/inspector_v2/widget_properties/properties_view.dart';
 import 'package:devtools_app_shared/ui.dart';
 import 'package:devtools_test/helpers.dart';
