+++ packages/devtools_app/test/inspector_v2/inspector_tree_test.dart
@@ -64,6 +64,7 @@ void main() {
         inspectorV2: inspectorController,
         debugger: debuggerController,
         InspectorTree(
+          controller: inspectorController,
           treeController: treeController,
         ),
       ),
