+++ packages/devtools_app/test/shared/visible_screens_test.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:devtools_app/devtools_app.dart';
+import 'package:devtools_app/src/screens/inspector_shared/inspector_screen.dart';
 import 'package:devtools_app/src/shared/framework_controller.dart';
 import 'package:devtools_app_shared/utils.dart';
 import 'package:devtools_shared/devtools_shared.dart';
